skills/csv-cleaner/requirements.txt
pandas>=2.0
chardet
python-dateutil
phonenumbers
jsonschema
